+++ src/ehrdata/io/__init__.py
@@ -1,5 +1,5 @@
 from . import omop
 from .csv import read_csv
-from .h5ad import read_h5ad, write_h5ad
+from .h5ed import read_h5ad, read_h5ed, write_h5ad, write_h5ed
 from .pandas import from_pandas, to_pandas
 from .zarr import read_zarr, write_zarr